Implement new events for Android (1/2)
