+++ android/src/main/java/com/mrousavy/camera/utils/CameraCharacteristicsUtils.kt
@@ -4,7 +4,6 @@ import android.hardware.camera2.CameraCharacteristics
 import android.util.Size
 import com.facebook.react.bridge.Arguments
 import com.facebook.react.bridge.ReadableArray
-import com.mrousavy.camera.parsers.bigger
 import kotlin.math.PI
 import kotlin.math.atan
